+++ server/src/test/java/org/opensearch/index/query/BoolQueryBuilderTests.java
@@ -497,4 +497,79 @@ public class BoolQueryBuilderTests extends AbstractQueryTestCase<BoolQueryBuilde
         assertEquals(0, set.size());
 
     }
+
+    public void testOneMustNotRangeRewritten() throws Exception {
+        int from = 10;
+        int to = 20;
+        for (boolean includeLower : new boolean[] { true, false }) {
+            for (boolean includeUpper : new boolean[] { true, false }) {
+                BoolQueryBuilder qb = new BoolQueryBuilder();
+                QueryBuilder rq = getRangeQueryBuilder(INT_FIELD_NAME, from, to, includeLower, includeUpper);
+                qb.mustNot(rq);
+
+                BoolQueryBuilder rewritten = (BoolQueryBuilder) Rewriteable.rewrite(qb, createShardContext());
+                assertFalse(rewritten.mustNot().contains(rq));
+
+                QueryBuilder expectedLowerQuery = getRangeQueryBuilder(INT_FIELD_NAME, null, from, false, !includeLower);
+                QueryBuilder expectedUpperQuery = getRangeQueryBuilder(INT_FIELD_NAME, to, null, !includeUpper, true);
+                assertEquals(1, rewritten.must().size());
+
+                BoolQueryBuilder nestedBoolQuery = (BoolQueryBuilder) rewritten.must().get(0);
+                assertEquals(2, nestedBoolQuery.should().size());
+                assertEquals("1", nestedBoolQuery.minimumShouldMatch());
+                assertTrue(nestedBoolQuery.should().contains(expectedLowerQuery));
+                assertTrue(nestedBoolQuery.should().contains(expectedUpperQuery));
+            }
+        }
+    }
+
+    public void testOneSingleEndedMustNotRangeRewritten() throws Exception {
+        // Test a must_not range query with only one endpoint is rewritten correctly
+        int from = 10;
+        BoolQueryBuilder qb = new BoolQueryBuilder();
+        QueryBuilder rq = getRangeQueryBuilder(INT_FIELD_NAME, from, null, false, false);
+        qb.mustNot(rq);
+        BoolQueryBuilder rewritten = (BoolQueryBuilder) Rewriteable.rewrite(qb, createShardContext());
+        assertFalse(rewritten.mustNot().contains(rq));
+
+        QueryBuilder expectedQuery = getRangeQueryBuilder(INT_FIELD_NAME, null, from, false, true);
+        assertEquals(1, rewritten.must().size());
+        BoolQueryBuilder nestedBoolQuery = (BoolQueryBuilder) rewritten.must().get(0);
+        assertEquals(1, nestedBoolQuery.should().size());
+        assertTrue(nestedBoolQuery.should().contains(expectedQuery));
+        assertEquals("1", nestedBoolQuery.minimumShouldMatch());
+    }
+
+    public void testMultipleMustNotRangesNotRewritten() throws Exception {
+        BoolQueryBuilder qb = new BoolQueryBuilder();
+        // Test a field with two ranges is not rewritten
+        QueryBuilder rq1of2 = new RangeQueryBuilder(INT_RANGE_FIELD_NAME).gt(10).lt(20);
+        QueryBuilder rq2of2 = new RangeQueryBuilder(INT_RANGE_FIELD_NAME).gt(30).lt(40);
+        qb.mustNot(rq1of2);
+        qb.mustNot(rq2of2);
+        BoolQueryBuilder rewritten = (BoolQueryBuilder) Rewriteable.rewrite(qb, createShardContext());
+
+        assertTrue(rewritten.mustNot().contains(rq1of2));
+        assertTrue(rewritten.mustNot().contains(rq2of2));
+        assertEquals(0, rewritten.should().size());
+    }
+
+    private QueryBuilder getRangeQueryBuilder(String fieldName, Integer lower, Integer upper, boolean includeLower, boolean includeUpper) {
+        RangeQueryBuilder rq = new RangeQueryBuilder(fieldName);
+        if (lower != null) {
+            if (includeLower) {
+                rq.gte(lower);
+            } else {
+                rq.gt(lower);
+            }
+        }
+        if (upper != null) {
+            if (includeUpper) {
+                rq.lte(upper);
+            } else {
+                rq.lt(upper);
+            }
+        }
+        return rq;
+    }
 }